Move `equal_ignoring_nullability()` method to `Type`.
